Fixes and removed a useless folder
- Removed ogg folder from framework folder
- Some tweaks to OpenALSoundManager
- Fixed Quantize on x64, (wrong type chosen in sizeof)
- Probably fixed the null musicinterface when reloading?

+++ src/SexyAppFramework/Quantize.cpp
@@ -62,7 +62,7 @@ bool Sexy::Quantize8Bit(
 
 				memmove(aTranslationTable + anInsertPos + 1,
 						aTranslationTable + anInsertPos,
-						(aColorTableSize - anInsertPos) * sizeof(uint32_t));
+						(aColorTableSize - anInsertPos) * sizeof(uint8_t));
 				aTranslationTable[anInsertPos] = aColorTableSize;
 
 				theDestColorTable[aColorTableSize] = aColor;
